moved methods that didn't use course model from courseService to discordBot service
