+++ src/discordBot/services/service.js
@@ -2,6 +2,7 @@ const axios = require("axios");
 const fs = require("fs");
 const path = require("path");
 const { logError } = require("./logger");
+const { courseAdminRole } = require("../../../config.json");
 
 require("dotenv").config();
 const GUIDE_CHANNEL_NAME = "guide";
@@ -299,6 +300,44 @@ const getUserWithUserId = async (guild, userId) => {
   return await guild.members.cache.get(userId);
 };
 
+const changeCourseRoles = async (courseName, newValue, guild) => {
+  await Promise.all(guild.roles.cache
+    .filter(r => (r.name === `${courseName} ${courseAdminRole}` || r.name === courseName))
+    .map(async role => {
+      if (role.name.includes("instructor")) {
+        role.setName(`${newValue} instructor`);
+      }
+      else {
+        role.setName(newValue);
+      }
+    },
+    ));
+};
+
+const changeInvitationLink = async (channelAnnouncement) => {
+  const pinnedMessages = await channelAnnouncement.messages.fetchPinned();
+  const invMessage = pinnedMessages.find(msg => msg.author.bot && msg.content.includes("Invitation link for the course"));
+  const courseName = channelAnnouncement.parent.name.replace(/([\uE000-\uF8FF]|\uD83C[\uDF00-\uDFFF]|\uD83D[\uDC00-\uDDFF])/g, "").trim();
+  const updatedMsg = createCourseInvitationLink(courseName);
+  await invMessage.edit(updatedMsg);
+};
+
+const setEmojisLock = async (category, hidden, courseName) => {
+  hidden ? await category.setName(`👻🔐 ${courseName}`) : await category.setName(`📚🔐 ${courseName}`);
+};
+
+const setEmojisUnlock = async (category, hidden, courseName) => {
+  hidden ? await category.setName(`👻 ${courseName}`) : await category.setName(`📚 ${courseName}`);
+};
+
+const setEmojisHide = async (category, locked, courseName) => {
+  locked ? await category.setName(`👻🔐 ${courseName}`) : await category.setName(`👻 ${courseName}`);
+};
+
+const setEmojisUnhide = async (category, locked, courseName) => {
+  locked ? await category.setName(`📚🔐 ${courseName}`) : await category.setName(`📚 ${courseName}`);
+};
+
 module.exports = {
   findCategoryWithCourseName,
   findOrCreateRoleWithName,
@@ -322,4 +361,10 @@ module.exports = {
   getCategoryChannelPermissionOverwrites,
   getDefaultChannelObjects,
   getCategoryObject,
+  changeCourseRoles,
+  changeInvitationLink,
+  setEmojisLock,
+  setEmojisUnlock,
+  setEmojisHide,
+  setEmojisUnhide,
 };
